requirements.txt
PyPDF2==3.0.1
Pillow==10.1.0
PyMuPDF==1.23.8
